2020/01/07 安丘 14点

+++ src/main/java/com/shenke/entity/YuanLiaoRequire.java
@@ -30,13 +30,13 @@ public class YuanLiaoRequire {
 
     private String state; //状态
 
-    private Integer sumNum;//总件数
+    private Double sumNum;//总件数
 
-    public Integer getSumNum() {
+    public Double getSumNum() {
         return sumNum;
     }
 
-    public void setSumNum(Integer sumNum) {
+    public void setSumNum(Double sumNum) {
         this.sumNum = sumNum;
     }
